+++ KursovMakarovAlex/texturesimport.h
@@ -32,4 +32,4 @@ SDL_Texture* CreateUTexture(const char filename[]);
 
 SDL_Texture* CaptureScreenTexture(SDL_Renderer* renderer);
 
-SDL_Rect GetTextureAllRect(SDL_Texture* texture, int scale);
+SDL_Rect GetTextureAllRect(SDL_Texture* texture, float scale);
